+++ src/types.ts
@@ -1,0 +1,9 @@
+export interface IConfig {
+  showImageTypes: string[],
+  keyword: string,
+  activeKey: string[]
+  backgroundColor: string,
+  size: number,
+  includeFolders: string[],
+  excludeFolders: string[],
+}