+++ src/webview/PreviewImages/index.tsx
@@ -1,4 +1,5 @@
 import {
+  Alert,
   Button,
   Checkbox,
   Collapse,
@@ -13,9 +14,9 @@ import {
   Tag,
   Tooltip
 } from 'antd'
-import { FolderOpenTwoTone, InfoCircleOutlined, SearchOutlined } from '@ant-design/icons'
+import { FolderOpenTwoTone, InfoCircleOutlined, SearchOutlined, SettingOutlined } from '@ant-design/icons'
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
-import { MESSAGE_CMD } from '../../constants'
+import { BACKGROUND_COLOR_OPTIONS, DEFAULT_BACKGROUND_COLOR, DEFAULT_IMAGE_SIZE, MESSAGE_CMD } from '../../constants'
 import { callVscode } from '@easy_vscode/webview'
 import ImageLazyLoad from './ImageLazyLoad'
 import {
@@ -25,6 +26,7 @@ import {
   StyledPicCount,
   StyledPreviewImages,
   StyledReloadOutlined,
+  StyledSettingOutlined,
   StyleImage,
   StyleImageList,
   StyleRowTitle,
@@ -34,6 +36,10 @@ import {
 import ImageInfo from './ImageInfo'
 import { useDebounceFn, useScroll } from 'ahooks'
 import { BUILTIN_MESSAGE_CMD } from '@easy_vscode/core/lib/constants'
+import { IConfig } from 'types'
+import SettingsModal from './SettingsModal'
+
+declare const window: any;
 
 const completeImgs = (imgs, projectPath) => {
   return imgs.map((img) => {
@@ -52,20 +58,7 @@ const completeImgs = (imgs, projectPath) => {
   })
 }
 
-const backgroundColorOptions = [
-  '#ffffff',
-  '#cccccc',
-  '#999999',
-  '#333333',
-  '#a89a89',
-  '#a9e4af',
-  '#f1a8a4',
-  '#64bbe2',
-  '#8488b6'
-]
-
 const THRESHOLD_ALL_COLLAPSED = 1200
-const DEFAULT_IMAGE_SIZE = 100
 const THRESHOLD_ENABLE_LAZY_LOADING = 150
 const THRESHOLD_DELAY_CHANGE_SIZE = 200
 
@@ -88,16 +81,19 @@ const PreviewImages: React.FC = () => {
   const [activeKey, setActiveKey] = useState<string[]>([])
   const [allPaths, setAllPaths] = useState<string[]>([])
   const [showImgs, setShowImgs] = useState<IImage[]>([])
-  const [backgroundColor, setBackgroundColor] = useState<string>(backgroundColorOptions[1])
+  const [backgroundColor, setBackgroundColor] = useState<string>(DEFAULT_BACKGROUND_COLOR)
   const [keyword, setKeyword] = useState<string>('')
   const [beforeFetch, setBeforeFetch] = useState(true)
   const [loading, setLoading] = useState(false)
   const [size, setSize] = useState<number>(DEFAULT_IMAGE_SIZE)
   const [isScrolling, setIsScrolling] = useState(false)
-  const [relativeDir, setRelativeDir] = useState('')
-  const initClickFilePath = (window as any).commandArgs?.[0]?.path || ''
-  const [clickFilePath, setClickFilePath] = useState(initClickFilePath)
+  const [relativeDir, setRelativeDir] = useState<string>('')
+  const initClickFilePath = window?.commandArgs?.[0]?.path || '';
+  const [clickFilePath, setClickFilePath] = useState<string>(initClickFilePath)
   const [everAutoPreview, setEverAutoPreview] = useState(false)
+  const [showSettingsModal, setShowSettingsModal] = useState(false)
+  const [includeFolders, setIncludeFolders] = useState<string[]>([])
+  const [excludeFolders, setExcludeFolders] = useState<string[]>([])
   const currentProjectPath = useRef('')
 
   const { run: onDebounceScroll } = useDebounceFn(
@@ -125,7 +121,7 @@ const PreviewImages: React.FC = () => {
     return path.substring(0, path.lastIndexOf('/') + 1)
   }
 
-  const refreshImgs = () => {
+  const refreshImgs = useCallback(() => {
     setLoading(true)
     callVscode({ cmd: MESSAGE_CMD.GET_ALL_IMGS }, ({ imgs, projectPath }: { imgs: IImage[], projectPath: string }) => {
       currentProjectPath.current = projectPath
@@ -143,8 +139,8 @@ const PreviewImages: React.FC = () => {
       setBeforeFetch(false)
       updateImgs(imgs)
     })
-  }
-  useEffect(refreshImgs, [])
+  }, [clickFilePath])
+  useEffect(refreshImgs, [refreshImgs])
 
   const onRevealWebview = useCallback((event) => {
     const message = event?.data
@@ -272,9 +268,70 @@ const PreviewImages: React.FC = () => {
     setEverAutoPreview(true)
   }
 
+  /**
+   * save to local config file
+   */
+  useEffect(() => {
+    callVscode({
+      cmd: MESSAGE_CMD.GET_CONFIG,
+    }, (data: IConfig) => {
+      setBackgroundColor(data.backgroundColor)
+      setSize(data.size)
+      setShowImageTypes(data.showImageTypes)
+      setKeyword(data.keyword)
+      setActiveKey(data.activeKey)
+      setIncludeFolders(data.includeFolders)
+      setExcludeFolders(data.excludeFolders)
+    })
+  }, [])
+
+  /**
+   * save to local config file
+   */
+  useEffect(() => {
+    callVscode({
+      cmd: MESSAGE_CMD.SAVE_CONFIG,
+      data: {
+        backgroundColor,
+        size,
+        showImageTypes,
+        keyword,
+        activeKey,
+      }
+    })
+  }, [showImageTypes, backgroundColor, size, activeKey, keyword])
+
+  /**
+   * save to local config file and refresh images
+   */
+  useEffect(() => {
+    callVscode({
+      cmd: MESSAGE_CMD.SAVE_CONFIG,
+      data: {
+        includeFolders,
+        excludeFolders
+      }
+    }, refreshImgs)
+  }, [includeFolders, excludeFolders, refreshImgs])
+
+  const handleClickSettings = () => {
+    setShowSettingsModal(true)
+  }
+
+  const handleApplySettings = (includeFolders: string[], excludeFolders: string[]) => {
+    setIncludeFolders(includeFolders)
+    setExcludeFolders(excludeFolders)
+  }
+
   return (
     <ConfigProvider renderEmpty={customizeRenderEmpty}>
       <Spin spinning={loading}>
+        <Alert closable message={
+          <div>
+            New features: ① Individual project settings are now stored in local files. ② Search now has options to include or exclude specific folders. &nbsp;&nbsp;
+            <a href='https://github.com/ZhangJian1713/vscode-image-viewer/issues' target='_blank' rel="noreferrer">Report issues</a>
+          </div>
+        } type="info" showIcon />
         <StyledPreviewImages style={{ padding: '20px' }}>
           <StyleTopRows>
             <Input
@@ -282,10 +339,11 @@ const PreviewImages: React.FC = () => {
               allowClear
               size='middle'
               placeholder='image path/name'
-              style={{ width: 'calc(100% - 32px)', marginRight: '8px' }}
+              style={{ width: 'calc(100% - 60px)', marginRight: '8px' }}
               value={keyword}
               onChange={(e) => setKeyword(e.target.value)}
             />
+            <StyledSettingOutlined onClick={handleClickSettings} />
             <StyledReloadOutlined onClick={refreshImgs} />
           </StyleTopRows>
           {/* Type */}
@@ -308,8 +366,9 @@ const PreviewImages: React.FC = () => {
           <StyleTopRows style={{ marginBottom: '6px' }}>
             <StyleRowTitle>Background:</StyleRowTitle>
             <span>
-              {backgroundColorOptions.map((color) => (
+              {BACKGROUND_COLOR_OPTIONS.map((color) => (
                 <StyleSquare
+                  key={color}
                   onClick={() => setBackgroundColor(color)}
                   isSelected={backgroundColor === color}
                   color={color}
@@ -332,19 +391,21 @@ const PreviewImages: React.FC = () => {
           </StyleTopRows>
           {/* Expand/Collapse All */}
           <StyleTopRows>
-            <Space>
-              <span style={{ color: '#bbb' }}>
-                Search result: <span style={{ color: '#333' }}>{showImgs.length}</span>
-              </span>
-              <Tooltip
-                placement='right'
-                title={`When there are more than ${THRESHOLD_ALL_COLLAPSED} images(after being filtered) being displayed, all directories are collapsed by default.`}
-              >
-                <InfoCircleOutlined style={{ fontSize: '16px', color: '#ccc' }} />
-              </Tooltip>
-              <Button onClick={() => setActiveKey([...allPaths])}>Expand All</Button>
-              <Button onClick={() => setActiveKey([])}>Collapse All</Button>
-            </Space>
+            <StyledBetweenWrapper>
+              <Space>
+                <span style={{ color: '#bbb' }}>
+                  Search result: <span style={{ color: '#333' }}>{showImgs.length}</span>
+                </span>
+                <Tooltip
+                  placement='right'
+                  title={`When there are more than ${THRESHOLD_ALL_COLLAPSED} images(after being filtered) being displayed, all directories are collapsed by default.`}
+                >
+                  <InfoCircleOutlined style={{ fontSize: '16px', color: '#ccc' }} />
+                </Tooltip>
+                <Button onClick={() => setActiveKey([...allPaths])}>Expand All</Button>
+                <Button onClick={() => setActiveKey([])}>Collapse All</Button>
+              </Space>
+            </StyledBetweenWrapper>
           </StyleTopRows>
           {relativeDir && (
             <StyleTopRows>
@@ -361,7 +422,7 @@ const PreviewImages: React.FC = () => {
                     <Collapse.Panel
                       header={
                         <span>
-                          {path}
+                          {path.replace(/^\/|\/$/g, '')}
                           <StyledPicCount>({showImgs.filter((img) => img.dirPath === path).length})</StyledPicCount>
                           <StyledFolderOpenTwoTone>
                             <FolderOpenTwoTone twoToneColor='#f4d057' onClick={(e) => handleClickOpenFolder(e, path)} />
@@ -399,6 +460,17 @@ const PreviewImages: React.FC = () => {
           </div>
         </StyledPreviewImages>
       </Spin>
+      {
+        showSettingsModal && (
+          <SettingsModal
+            includeFolders={includeFolders}
+            excludeFolders={excludeFolders}
+            visible={showSettingsModal}
+            onClose={() => setShowSettingsModal(false)}
+            onApply={handleApplySettings}
+          />
+        )
+      }
     </ConfigProvider>
   )
 }