+++ src/webviewController/imagesViewer/index.ts
@@ -4,6 +4,7 @@ import { utils, webviewUtils } from '@easy_vscode/core'
 import { IWebview, IWebviewProps, IMessage } from '@easy_vscode/core/lib/types'
 import { DIST_WEBVIEW_INDEX_HTML, EXTENSION_COMMANDS, MESSAGE_CMD, WEBVIEW_NAMES } from '../../constants'
 import { getAllImgs, getImageBase64, getImageSize } from './utils'
+import { readLocalConfigFile, writeLocalConfigFile } from './config'
 
 const { deleteFile, getProjectPath, renameFile } = utils
 const { invokeCallback, successResp } = webviewUtils
@@ -66,7 +67,18 @@ const messageHandlers = new Map([
       const dimensions = getImageSize(message.data.filePath)
       invokeCallback(viewType, message, dimensions)
     }
-  ]
+  ],
+  [
+    MESSAGE_CMD.SAVE_CONFIG,
+    (message: IMessage) => {
+      writeLocalConfigFile(message.data)
+      invokeCallback(viewType, message, successResp)
+    }
+  ],
+  [
+    MESSAGE_CMD.GET_CONFIG,
+    (message: IMessage) => invokeCallback(viewType, message, readLocalConfigFile())
+  ],
 ])
 
 const webview: IWebview = { webviewProps, messageHandlers }